eyomn_conn_api/requirements.txt
fastapi[standard]
python-dotenv
slowapi
typing-extensions
uvicorn
python-multipart
pydantic
